+++ lib/assent/jwt_adapter.ex
@@ -37,8 +37,17 @@ defmodule Assent.JWTAdapter do
   @callback verify(binary(), binary() | map() | nil, Keyword.t()) ::
               {:ok, map()} | {:error, term()}
 
+  @default_jwt_adapter Assent.JWTAdapter.AssentJWT
+
   @doc """
-  Generates a signed JSON Web Token signature
+  Generates a signed JSON Web Token signature.
+
+  ## Options
+
+  - `:json_library` - The JSON library to use, optional, see
+    `Assent.Config.json_library/1`.
+  - `:jwt_adapter` - The JWT adapter module to use, optional, defaults to
+    `#{inspect(@default_jwt_adapter)}`
   """
   @spec sign(map(), binary(), binary(), Keyword.t()) :: {:ok, binary()} | {:error, term()}
   def sign(claims, alg, secret, opts \\ []) do
@@ -47,7 +56,14 @@ defmodule Assent.JWTAdapter do
   end
 
   @doc """
-  Verifies the JSON Web Token signature
+  Verifies the JSON Web Token signature.
+
+  ## Options
+
+  - `:json_library` - The JSON library to use, optional, see
+    `Assent.Config.json_library/1`.
+  - `:jwt_adapter` - The JWT adapter module to use, optional, defaults to
+    `#{inspect(@default_jwt_adapter)}`
   """
   @spec verify(binary(), binary() | map() | nil, Keyword.t()) :: {:ok, map()} | {:error, any()}
   def verify(token, secret, opts \\ []) do
@@ -57,7 +73,7 @@ defmodule Assent.JWTAdapter do
 
   defp get_adapter(opts) do
     default_opts = Keyword.put(opts, :json_library, Config.json_library(opts))
-    default_jwt_adapter = Application.get_env(:assent, :jwt_adapter, Assent.JWTAdapter.AssentJWT)
+    default_jwt_adapter = Application.get_env(:assent, :jwt_adapter, @default_jwt_adapter)
 
     case Keyword.get(opts, :jwt_adapter, default_jwt_adapter) do
       {adapter, opts} -> {adapter, Keyword.merge(default_opts, opts)}
@@ -66,9 +82,14 @@ defmodule Assent.JWTAdapter do
   end
 
   @doc """
-  Loads a private key from the provided configuration
+  Loads a private key from the provided configuration.
+
+  ## Options
+
+  - `:private_key_path` - The path to the private key file, optional.
+  - `:private_key` - The private key, required if `:private_key_path` is not set.
   """
-  @spec load_private_key(Config.t()) :: {:ok, binary()} | {:error, term()}
+  @spec load_private_key(Keyword.t()) :: {:ok, binary()} | {:error, term()}
   def load_private_key(config) do
     case Config.fetch(config, :private_key_path) do
       {:ok, path} -> read(path)