+++ test/assent/strategy_test.exs
@@ -2,271 +2,48 @@ defmodule Assent.StrategyTest do
   use Assent.TestCase
   doctest Assent.Strategy
 
-  alias Assent.{HTTPAdapter.HTTPResponse, InvalidResponseError, Strategy}
+  alias Assent.Strategy
 
-  @body %{"a" => "1", "b" => "2"}
-  @headers [{"content-type", "application/json"}]
-  @json_encoded_body @json_library.encode!(@body)
-  @uri_encoded_body URI.encode_query(@body)
-
-  test "decode_response/2" do
-    assert {:ok, response} =
-             Strategy.decode_response(
-               %HTTPResponse{body: @json_encoded_body, headers: @headers},
-               []
-             )
-
-    assert response.body == @body
-
-    assert {:ok, response} =
-             Strategy.decode_response(
-               %HTTPResponse{
-                 body: @json_encoded_body,
-                 headers: [{"content-type", "application/json; charset=utf-8"}]
-               },
-               []
-             )
-
-    assert response.body == @body
-
-    assert {:ok, response} =
-             Strategy.decode_response(
-               %HTTPResponse{
-                 body: @json_encoded_body,
-                 headers: [{"content-type", "text/javascript"}]
-               },
-               []
-             )
-
-    assert response.body == @body
-
-    assert {:ok, response} =
-             Strategy.decode_response(
-               %HTTPResponse{
-                 body: @uri_encoded_body,
-                 headers: [{"content-type", "application/x-www-form-urlencoded"}]
-               },
-               []
-             )
-
-    assert response.body == @body
-
-    assert {:ok, response} = Strategy.decode_response(%HTTPResponse{body: @body, headers: []}, [])
-    assert response.body == @body
-
-    assert {:error, %InvalidResponseError{} = error} =
-             Strategy.decode_response(%HTTPResponse{body: "%", headers: @headers}, [])
-
-    assert error.response.body == "%"
+  defmodule HTTPMock do
+    def request(:get, _, _, _, _), do: {:error, __MODULE__}
   end
 
   defmodule JSONMock do
     def decode(_string), do: {:ok, :decoded}
   end
 
+  defmodule JWTMock do
+    @moduledoc false
+
+    def sign(_claims, _alg, _secret, opts), do: {:error, opts}
+
+    def verify(_binary, _secret, opts), do: {:error, opts}
+  end
+
+  test "http_request/5" do
+    config = [http_adapter: HTTPMock, json_library: JSONMock]
+
+    assert {:error, error} = Strategy.http_request(:get, "/path", nil, [], config)
+    assert error.reason == HTTPMock
+  end
+
+  test "sign_jwt/4" do
+    config = [json_library: JSONMock, jwt_adapter: JWTMock, private_key: "myprivatekey.pem", a: 1]
+
+    assert {:error, opts} = Strategy.sign_jwt(%{"claim" => "a"}, "alg", "secret", config)
+    assert opts == [json_library: JSONMock, jwt_adapter: JWTMock]
+  end
+
+  test "verify_jwt/4" do
+    config = [json_library: JSONMock, jwt_adapter: JWTMock, private_key: "myprivatekey.pem", a: 1]
+
+    assert {:error, opts} = Strategy.verify_jwt("token", "secret", config)
+    assert opts == [json_library: JSONMock, jwt_adapter: JWTMock]
+  end
+
   test "decode_json/2" do
     assert Strategy.decode_json("{\"a\": 1}", []) == {:ok, %{"a" => 1}}
     assert Strategy.decode_json("{\"a\": 1}", json_library: JSONMock) == {:ok, :decoded}
-  end
-
-  defmodule HTTPMock do
-    @json_library (Code.ensure_loaded?(JSON) && JSON) || Jason
-
-    def request(:get, "http-adapter", nil, [], nil) do
-      {:ok, %HTTPResponse{status: 200, headers: [], body: nil}}
-    end
-
-    def request(:get, "http-adapter-with-opts", nil, [], opts) do
-      {:ok, %HTTPResponse{status: 200, headers: [], body: opts}}
-    end
-
-    def request(:get, "json-encoded-body", nil, [], nil) do
-      {:ok,
-       %HTTPResponse{
-         status: 200,
-         headers: [{"content-type", "application/json"}],
-         body: @json_library.encode!(%{"a" => 1})
-       }}
-    end
-
-    def request(:get, "json-encoded-body-already-decoded", nil, [], nil) do
-      {:ok,
-       %HTTPResponse{
-         status: 200,
-         headers: [{"content-type", "application/json"}],
-         body: %{"a" => 1}
-       }}
-    end
-
-    def request(:get, "json-encoded-body-text/javascript-header", nil, [], nil) do
-      {:ok,
-       %HTTPResponse{
-         status: 200,
-         headers: [{"content-type", "text/javascript"}],
-         body: @json_library.encode!(%{"a" => 1})
-       }}
-    end
-
-    def request(:get, "invalid-json-body", nil, [], nil) do
-      {:ok,
-       %HTTPResponse{status: 200, headers: [{"content-type", "application/json"}], body: "%"}}
-    end
-
-    def request(:get, "json-no-headers", nil, [], nil) do
-      {:ok, %HTTPResponse{status: 200, headers: [], body: @json_library.encode!(%{"a" => 1})}}
-    end
-
-    def request(:get, "form-data-body", nil, [], nil) do
-      {:ok,
-       %HTTPResponse{
-         status: 200,
-         headers: [{"content-type", "application/x-www-form-urlencoded"}],
-         body: URI.encode_query(%{"a" => 1})
-       }}
-    end
-
-    def request(:get, "form-data-body-already-decoded", nil, [], nil) do
-      {:ok,
-       %HTTPResponse{
-         status: 200,
-         headers: [{"content-type", "application/x-www-form-urlencoded"}],
-         body: %{"a" => 1}
-       }}
-    end
-  end
-
-  test "request/5" do
-    assert Strategy.request(:get, "http-adapter", nil, [], http_adapter: HTTPMock) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [],
-                body: nil,
-                http_adapter: HTTPMock,
-                request_url: "http-adapter"
-              }}
-
-    assert Strategy.request(:get, "http-adapter-with-opts", nil, [],
-             http_adapter: {HTTPMock, a: 1}
-           ) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [],
-                body: [a: 1],
-                http_adapter: HTTPMock,
-                request_url: "http-adapter-with-opts"
-              }}
-
-    assert Strategy.request(:get, "json-encoded-body", nil, [], http_adapter: HTTPMock) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [{"content-type", "application/json"}],
-                body: %{"a" => 1},
-                http_adapter: HTTPMock,
-                request_url: "json-encoded-body"
-              }}
-
-    assert Strategy.request(:get, "json-encoded-body-already-decoded", nil, [],
-             http_adapter: HTTPMock
-           ) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [{"content-type", "application/json"}],
-                body: %{"a" => 1},
-                http_adapter: HTTPMock,
-                request_url: "json-encoded-body-already-decoded"
-              }}
-
-    assert Strategy.request(:get, "json-encoded-body-text/javascript-header", nil, [],
-             http_adapter: HTTPMock
-           ) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [{"content-type", "text/javascript"}],
-                body: %{"a" => 1},
-                http_adapter: HTTPMock,
-                request_url: "json-encoded-body-text/javascript-header"
-              }}
-
-    assert {:error, %InvalidResponseError{}} =
-             Strategy.request(:get, "invalid-json-body", nil, [], http_adapter: HTTPMock)
-
-    assert Strategy.request(:get, "json-no-headers", nil, [], http_adapter: HTTPMock) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [],
-                body: @json_library.encode!(%{"a" => 1}),
-                http_adapter: HTTPMock,
-                request_url: "json-no-headers"
-              }}
-
-    assert Strategy.request(:get, "form-data-body", nil, [], http_adapter: HTTPMock) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [{"content-type", "application/x-www-form-urlencoded"}],
-                body: %{"a" => "1"},
-                http_adapter: HTTPMock,
-                request_url: "form-data-body"
-              }}
-
-    assert Strategy.request(:get, "form-data-body-already-decoded", nil, [],
-             http_adapter: HTTPMock
-           ) ==
-             {:ok,
-              %HTTPResponse{
-                status: 200,
-                headers: [{"content-type", "application/x-www-form-urlencoded"}],
-                body: %{"a" => 1},
-                http_adapter: HTTPMock,
-                request_url: "form-data-body-already-decoded"
-              }}
-  end
-
-  defmodule CustomJWTAdapter do
-    @moduledoc false
-
-    def sign(_claims, _alg, _secret, _opts), do: :signed
-
-    def verify(_binary, _secret, _opts), do: :verified
-  end
-
-  defmodule CustomJSONLibrary do
-    @moduledoc false
-
-    def decode(_binary), do: {:ok, %{"alg" => "none", "custom_json" => true}}
-
-    def encode!(_any), do: ""
-  end
-
-  @claims %{"iat" => 1_516_239_022, "name" => "John Doe", "sub" => "1234567890"}
-  @alg "HS256"
-  @secret "your-256-bit-secret"
-  @token "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE1MTYyMzkwMjIsIm5hbWUiOiJKb2huIERvZSIsInN1YiI6IjEyMzQ1Njc4OTAifQ.fdOPQ05ZfRhkST2-rIWgUpbqUsVhkkNVNcuG7Ki0s-8"
-
-  @empty_encoding Base.url_encode64("", padding: false)
-
-  test "sign_jwt/2" do
-    assert Strategy.sign_jwt(@claims, @alg, @secret, []) == {:ok, @token}
-
-    assert Strategy.sign_jwt(@token, @alg, @secret, jwt_adapter: CustomJWTAdapter) == :signed
-
-    assert {:ok, @empty_encoding <> "." <> _rest} =
-             Strategy.sign_jwt(@token, @alg, @secret, json_library: CustomJSONLibrary)
-  end
-
-  test "verify_jwt/2" do
-    assert {:ok, jwt} = Strategy.verify_jwt(@token, @secret, [])
-    assert jwt.verified?
-    assert Strategy.verify_jwt(@token, @secret, jwt_adapter: CustomJWTAdapter) == :verified
-
-    assert {:ok, %{header: %{"custom_json" => true}}} =
-             Strategy.verify_jwt(@token, @secret, json_library: CustomJSONLibrary)
   end
 
   test "to_url/3" do